Extract `ActionButtonView` component
