+++ frontend/src/metabase/writeback/components/ActionButtonViz/ActionButton.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useMemo } from "react";
 import type { Dashboard } from "metabase-types/api";
 import type { VisualizationProps } from "metabase-types/types/Visualization";
 
-import { StyledButton } from "./ActionButton.styled";
+import ActionButtonView from "./ActionButtonView";
 
 interface ActionButtonProps extends VisualizationProps {
   dashboard: Dashboard;
@@ -15,14 +15,6 @@ function ActionButton({
   getExtraDataForClick,
   onVisualizationClick,
 }: ActionButtonProps) {
-  const label = settings["button.label"];
-  const variant = settings["button.variant"];
-
-  const variantProps: any = {};
-  if (variant !== "default") {
-    variantProps[variant] = true;
-  }
-
   const clickObject = useMemo(() => ({ settings }), [settings]);
 
   const extraData = useMemo(
@@ -38,13 +30,11 @@ function ActionButton({
   }, [clickObject, extraData, onVisualizationClick]);
 
   return (
-    <StyledButton
+    <ActionButtonView
       onClick={onClick}
+      settings={settings}
       isFullHeight={!isSettings}
-      {...variantProps}
-    >
-      {label}
-    </StyledButton>
+    />
   );
 }
